+++ model/src/main/kotlin/io/spine/validation/ValidationPlugin.kt
@@ -29,6 +29,10 @@ package io.spine.validation
 import io.spine.protodata.plugin.Plugin
 import io.spine.protodata.plugin.Policy
 import io.spine.protodata.plugin.ViewRepository
+import io.spine.validation.required.RequiredFieldRepository
+import io.spine.validation.required.RequiredIdOptionPolicy
+import io.spine.validation.required.RequiredIdPatternPolicy
+import io.spine.validation.required.RequiredPolicy
 
 /**
  * A ProtoData plugin which attaches validation-related policies and views.
